♻️ slug api call
